requirements.txt
Faker==19.3.0
greenlet==2.0.2
psycopg2==2.9.7
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==1.4.17
